+++ src/components/mod.rs
@@ -1,0 +1,2 @@
+pub mod favourite_animal;
+pub mod incr_button;
