+++ src/lib.rs
@@ -3,6 +3,7 @@ use leptos_meta::*;
 use leptos_router::*;
 
 // Modules
+mod components;
 mod pages;
 
 // Top-Level pages
